[FLINK-25744] Support native savepoints
We introduce a savepoint type flag both to the CLI and REST API that
controls the binary format of the savepoint to take. Native savepoints
are taken in the state backend specific binary format. They can be
faster to take and restore from, but they do not support e.g. changing
the state backend.

This commit does not support RocksDB incremental savepoints yet. When
incremental savepoints are configured, native savepoints will fail.